Simplification
I simplified a lot, and generalized by the way
