+++ database/factories/SampleSiteFactory.php
@@ -12,11 +12,17 @@
 */
 
 $factory->define(App\Models\SampleSite::class, function (Faker\Generator $faker) {
+    $station_id = $faker->numberBetween(1, 50);
+    $code = App\Models\SampleSite::select('code')->whereStationId($station_id)->orderBy('code', 'dsc')->first();
+    if (!isset($code)) {
+        $code = 0;
+    }
+    $code++;
     return [
         'name' => strtoupper($faker->word),
-        'code' => $faker->unique()->randomNumber(3),
+        'code' => $faker->unique()->numberBetween(1, 50),
         'x' => $faker->randomFloat(3, -999, 999),
         'y' => $faker->randomFloat(3, -999, 999),
-        'station_id' => $faker->numberBetween(1, 50),
+        'station_id' => $station_id,
     ];
 });
